Move batching tests to separate file
