+++ server/jetstream_test.go
@@ -21050,191 +21050,3 @@ func TestJetStreamAllowMsgCounterSourceStartingAboveZero(t *testing.T) {
 	t.Run("R1", func(t *testing.T) { test(t, 1) })
 	t.Run("R3", func(t *testing.T) { test(t, 3) })
 }
-
-func TestJetStreamAtomicBatchPublish(t *testing.T) {
-	test := func(
-		t *testing.T,
-		nc *nats.Conn,
-		js nats.JetStreamContext,
-		storage StorageType,
-		retention RetentionPolicy,
-		replicas int,
-	) {
-		var pubAck JSPubAckResponse
-
-		cfg := &StreamConfig{
-			Name:      "TEST",
-			Subjects:  []string{"foo.*"},
-			Storage:   storage,
-			Retention: retention,
-			Replicas:  replicas,
-		}
-
-		_, err := jsStreamCreate(t, nc, cfg)
-		require_NoError(t, err)
-
-		m := nats.NewMsg("foo.0")
-		m.Data = []byte("foo.0")
-		m.Header.Set("Nats-Batch-Id", "uuid")
-
-		// Publish with atomic publish disabled.
-		rmsg, err := nc.RequestMsg(m, time.Second)
-		require_NoError(t, err)
-		require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-		require_NotNil(t, pubAck.Error)
-		require_Error(t, pubAck.Error, NewJSAtomicPublishDisabledError())
-
-		// Enable atomic publish.
-		cfg.AllowAtomicPublish = true
-		_, err = jsStreamUpdate(t, nc, cfg)
-		require_NoError(t, err)
-
-		// Publish without batch sequence errors.
-		rmsg, err = nc.RequestMsg(m, time.Second)
-		require_NoError(t, err)
-		pubAck = JSPubAckResponse{}
-		require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-		require_Error(t, pubAck.Error, NewJSAtomicPublishMissingSeqError())
-
-		// Publish a batch, misses start.
-		m.Header.Set("Nats-Batch-Id", "uuid")
-		m.Header.Set("Nats-Batch-Sequence", "2")
-		rmsg, err = nc.RequestMsg(m, time.Second)
-		require_NoError(t, err)
-		pubAck = JSPubAckResponse{}
-		require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-		require_Error(t, pubAck.Error, NewJSAtomicPublishIncompleteBatchError())
-
-		// Publish a "batch" which immediately commits.
-		m.Header.Set("Nats-Batch-Id", "uuid")
-		m.Header.Set("Nats-Batch-Sequence", "1")
-		m.Header.Set("Nats-Batch-Commit", "1")
-		rmsg, err = nc.RequestMsg(m, time.Second)
-		require_NoError(t, err)
-		pubAck = JSPubAckResponse{}
-		require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-		require_Equal(t, pubAck.Sequence, 1)
-		require_Equal(t, pubAck.BatchId, "uuid")
-		require_Equal(t, pubAck.BatchSize, 1)
-
-		// Reset commit.
-		m.Header.Del("Nats-Batch-Commit")
-
-		// Publish a "batch" which has gaps.
-		require_NoError(t, nc.PublishMsg(m))
-		m.Header.Set("Nats-Batch-Sequence", "3")
-		rmsg, err = nc.RequestMsg(m, time.Second)
-		require_NoError(t, err)
-		pubAck = JSPubAckResponse{}
-		require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-		require_Error(t, pubAck.Error, NewJSAtomicPublishIncompleteBatchError())
-
-		// Publish a batch of N messages.
-		m.Header.Del("Nats-Batch-Commit")
-		for seq, batch := uint64(1), uint64(5); seq <= batch; seq++ {
-			m.Subject = fmt.Sprintf("foo.%d", seq)
-			m.Data = []byte(m.Subject)
-			m.Header.Set("Nats-Batch-Sequence", strconv.FormatUint(seq, 10))
-			// If not commit.
-			if seq != batch {
-				require_NoError(t, nc.PublishMsg(m))
-				continue
-			}
-
-			m.Header.Set("Nats-Batch-Commit", "1")
-			rmsg, err = nc.RequestMsg(m, time.Second)
-			require_NoError(t, err)
-
-			pubAck = JSPubAckResponse{}
-			require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-			require_Equal(t, pubAck.Sequence, 6)
-			require_Equal(t, pubAck.BatchId, "uuid")
-			require_Equal(t, pubAck.BatchSize, 5)
-		}
-
-		// Validate stream contents.
-		if retention != InterestPolicy {
-			for i := 0; i < 6; i++ {
-				rsm, err := js.GetMsg("TEST", uint64(i+1))
-				require_NoError(t, err)
-				subj := fmt.Sprintf("foo.%d", i)
-				require_Equal(t, rsm.Subject, subj)
-				require_Equal(t, string(rsm.Data), subj)
-			}
-		}
-
-		// TODO(mvv): implement timeout
-	}
-
-	for _, storage := range []StorageType{FileStorage, MemoryStorage} {
-		for _, retention := range []RetentionPolicy{LimitsPolicy, InterestPolicy, WorkQueuePolicy} {
-			replicas := 3
-			t.Run(fmt.Sprintf("%s/%s/R%d", storage, retention, replicas), func(t *testing.T) {
-				c := createJetStreamClusterExplicit(t, "R3S", 3)
-				defer c.shutdown()
-
-				nc, js := jsClientConnect(t, c.randomServer())
-				defer nc.Close()
-
-				test(t, nc, js, storage, retention, replicas)
-			})
-		}
-	}
-}
-
-func TestJetStreamAtomicBatchPublishDedupeNotAllowed(t *testing.T) {
-	test := func(
-		t *testing.T,
-		nc *nats.Conn,
-		js nats.JetStreamContext,
-		storage StorageType,
-		retention RetentionPolicy,
-		replicas int,
-	) {
-		cfg := &StreamConfig{
-			Name:               "TEST",
-			Retention:          retention,
-			Subjects:           []string{"foo"},
-			Replicas:           3,
-			Storage:            FileStorage,
-			AllowAtomicPublish: true,
-		}
-
-		_, err := jsStreamCreate(t, nc, cfg)
-		require_NoError(t, err)
-
-		m := nats.NewMsg("foo")
-		m.Header.Set("Nats-Batch-Id", "uuid")
-		m.Header.Set("Nats-Batch-Sequence", "1")
-		m.Header.Set("Nats-Msg-Id", "msgId1")
-		require_NoError(t, nc.PublishMsg(m))
-		m.Header.Set("Nats-Batch-Sequence", "2")
-		m.Header.Set("Nats-Msg-Id", "pre-existing")
-		require_NoError(t, nc.PublishMsg(m))
-
-		var pubAck JSPubAckResponse
-		m.Header.Set("Nats-Batch-Sequence", "3")
-		m.Header.Set("Nats-Msg-Id", "msgId2")
-		m.Header.Set("Nats-Batch-Commit", "1")
-		rmsg, err := nc.RequestMsg(m, time.Second)
-		require_NoError(t, err)
-		require_NoError(t, json.Unmarshal(rmsg.Data, &pubAck))
-		require_NotNil(t, pubAck.Error)
-		require_Error(t, pubAck.Error, NewJSAtomicPublishDuplicateError())
-	}
-
-	for _, storage := range []StorageType{FileStorage, MemoryStorage} {
-		for _, retention := range []RetentionPolicy{LimitsPolicy, InterestPolicy, WorkQueuePolicy} {
-			replicas := 3
-			t.Run(fmt.Sprintf("%s/%s/R%d", storage, retention, replicas), func(t *testing.T) {
-				c := createJetStreamClusterExplicit(t, "R3S", 3)
-				defer c.shutdown()
-
-				nc, js := jsClientConnect(t, c.randomServer())
-				defer nc.Close()
-
-				test(t, nc, js, storage, retention, replicas)
-			})
-		}
-	}
-}
